+++ launcher/gradle-plugin/src/main/kotlin/org/sourcegrade/jagr/gradle/task/submission/PrepareSubmissionTask.kt
@@ -1,0 +1,4 @@
+package org.sourcegrade.jagr.gradle.task.submission
+
+abstract class PrepareSubmissionTask {
+}
