+++ launcher/gradle-plugin/src/main/kotlin/org/sourcegrade/jagr/gradle/task/submission/SubmissionWriteInfoTask.kt
@@ -1,0 +1,7 @@
+package org.sourcegrade.jagr.gradle.task.submission
+
+import org.gradle.api.DefaultTask
+
+abstract class SubmissionWriteInfoTask : DefaultTask() {
+
+}
